improded documentation, reformatted code
