+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nakadion"
-version = "0.3.0"
+version = "0.3.1"
 authors = ["Christian Douven <chridou@users.noreply.github.com>"]
 readme = "README.md"
 license = "Apache-2.0/MIT"
