+++ src/connector.rs
@@ -89,15 +89,16 @@ pub struct ConnectorSettings {
     /// Maximum time in seconds a stream will live before connection is closed by the server.
     ///
     /// If 0 or unspecified will stream indefinitely.
-    /// If this timeout is reached, any pending messages (in the sense of `stream_limit`) will be flushed
-    /// to the client.
+    /// If this timeout is reached, any pending messages (in the sense of `stream_limit`)
+    /// will be flushed to the client.
     /// Stream initialization will fail if `stream_timeout` is lower than `batch_flush_timeout`.
     #[builder(default="Duration::from_secs(0)")]
     pub stream_timeout: Duration,
     /// Maximum time in seconds to wait for the flushing of each chunk (per partition).
     ///
-    ///  * If the amount of buffered Events reaches `batch_limit` before this `batch_flush_timeout`
-    ///  is reached, the messages are immediately flushed to the client and batch flush timer is reset.
+    ///  * If the amount of buffered Events reaches `batch_limit`
+    /// before this `batch_flush_timeout` is reached, the messages are immediately
+    /// flushed to the client and batch flush timer is reset.
     ///  * If 0 or undefined, will assume 30 seconds.
     #[builder(default="Duration::from_secs(0)")]
     pub batch_flush_timeout: Duration,
@@ -107,8 +108,8 @@ pub struct ConnectorSettings {
     ///  `batch_flush_timeout`.
     #[builder(default="0")]
     pub batch_limit: usize,
-    /// The amount of uncommitted events Nakadi will stream before pausing the stream. When in paused
-    /// state and commit comes - the stream will resume. Minimal value is 1.
+    /// The amount of uncommitted events Nakadi will stream before pausing the stream.
+    /// When in paused state and commit comes - the stream will resume. Minimal value is 1.
     #[builder(default="0")]
     pub max_uncommitted_events: usize,
     /// The URI prefix for the Nakadi Host, e.g. "https://my.nakadi.com"
@@ -169,13 +170,14 @@ impl ConnectorSettingsBuilder {
                    default.");
             builder
         };
-        let builder = if let Some(anv_val) = env::var("NAKADION_BATCH_LIMIT").ok() {
-            builder.batch_limit(anv_val.parse()
+        let builder =
+            if let Some(anv_val) = env::var("NAKADION_BATCH_LIMIT").ok() {
+                builder.batch_limit(anv_val.parse()
                 .map_err(|err| format!("Could not parse 'NAKADION_BATCH_LIMIT': {}", err))?)
-        } else {
-            warn!("Environment variable 'NAKADION_BATCH_LIMIT' not found. Using default.");
-            builder
-        };
+            } else {
+                warn!("Environment variable 'NAKADION_BATCH_LIMIT' not found. Using default.");
+                builder
+            };
         let builder = if let Some(anv_val) = env::var("NAKADION_MAX_UNCOMMITED_EVENTS").ok() {
             builder.max_uncommitted_events(anv_val.parse()
                     .map_err(|err| {
@@ -186,14 +188,15 @@ impl ConnectorSettingsBuilder {
                    default.");
             builder
         };
-        let builder = if let Some(anv_val) = env::var("NAKADION_NAKADI_HOST").ok() {
-            builder.nakadi_host(anv_val.parse()
+        let builder =
+            if let Some(anv_val) = env::var("NAKADION_NAKADI_HOST").ok() {
+                builder.nakadi_host(anv_val.parse()
                 .map_err(|err| format!("Could not parse 'NAKADION_NAKADI_HOST': {}", err))?)
-        } else {
-            warn!("Environment variable 'NAKADION_NAKADI_HOST' not found. It will have to be set \
-                   manually.");
-            builder
-        };
+            } else {
+                warn!("Environment variable 'NAKADION_NAKADI_HOST' not found. \
+                       It will have to be set manually.");
+                builder
+            };
         Ok(builder)
     }
 }
@@ -253,9 +256,12 @@ impl HyperClientConnector {
     pub fn from_env_with_client(client: Client,
                                 token_provider: BoxedTokenProvider)
                                 -> Result<HyperClientConnector, String> {
-        let builder = ConnectorSettingsBuilder::from_env().map_err(|err| format!("Could not create settings builder: {}", err))?;
-        let settings = builder.build()
-            .map_err(|err| format!("Could not create settings from builder: {}", err))?;
+        let builder = ConnectorSettingsBuilder::from_env()
+            .map_err(|err| format!("Could not create settings builder: {}", err))?;
+        let settings =
+            builder
+                .build()
+                .map_err(|err| format!("Could not create settings from builder: {}", err))?;
         info!("Creating HyperClientConnector from: {:?}", settings);
         Ok(HyperClientConnector::with_client_and_settings(client, token_provider, settings))
     }
@@ -322,14 +328,15 @@ impl ReadsStream for HyperClientConnector {
             Ok(mut rsp) => {
                 match rsp.status {
                     StatusCode::Ok => {
-                        let stream_id = if let Some(stream_id) = rsp.headers
-                            .get::<XNakadiStreamId>()
-                            .map(|v| StreamId(v.to_string())) {
+                        let stream_id = if let Some(stream_id) =
+                            rsp.headers
+                                .get::<XNakadiStreamId>()
+                                .map(|v| StreamId(v.to_string())) {
                             stream_id
                         } else {
                             bail!(ClientErrorKind::InvalidResponse("The response lacked the \
                                                                     'X-Nakadi-StreamId' header."
-                                .to_string()))
+                                                                           .to_string()))
                         };
                         Ok((rsp, stream_id))
                     }
@@ -441,11 +448,12 @@ impl ProvidesStreamInfo for HyperClientConnector {
             Ok(mut rsp) => {
                 match rsp.status {
                     StatusCode::Ok => {
-                        let payload: StreamInfo = serde_json::from_reader(rsp).map_err(|err| {
-                                ClientErrorKind::InvalidResponse(format!("Could not parse \
+                        let payload: StreamInfo = serde_json::from_reader(rsp)
+                            .map_err(|err| {
+                                         ClientErrorKind::InvalidResponse(format!("Could not parse \
                                                                           stream stats: {}",
-                                                                         err))
-                            })?;
+                                                                                  err))
+                                     })?;
                         Ok(payload)
                     }
                     StatusCode::BadRequest => {
