+++ src/lib.rs
@@ -1,39 +1,42 @@
 //! # Nakadion
-//! 
+//!
 //! A client for the [Nakadi](https://github.com/zalando/nakadi) Event Broker.
-//! 
+//!
 //! Nakadion uses the Subscription API of Nakadi.
 //!
 //! ## Consuming
-//! 
+//!
 //! Nakadion supports two modes of consuming events. A sequuential one and a cuncurrent one.
-//! 
+//!
 //! ### Sequential Consumption
-//! 
-//! In this mode Nakadion will read a batch from Nakadi then call a handler on theese 
+//!
+//! In this mode Nakadion will read a batch from Nakadi then call a handler on theese
 //! and afterwards try to commit the batch.
 //! This mode of operation is simple, straight forward and should be sufficient for most scenarios.
-//! 
+//! No batches are buffered when consuming sequentially.
+//!
 //! ### Concurrent Consumption
 //!
 //! ** DO NOT USE FOR NOW **
-//! 
-//! In this mode Nakadion will spawn a number of worker threads and distribute work among them based on 
-//! the `partion id` of a batch. The workers are not dedidacted to a partition. 
-//! Work is rather distributed based on a hash of the `partition id`. 
-//! 
+//!
+//! In this mode Nakadion will spawn a number of worker threads
+//! and distribute work among them based on
+//! the `partion id` of a batch. The workers are not dedidacted to a partition.
+//! Work is rather distributed based on a hash of the `partition id`.
+//!
 //! ## Performance
-//! 
+//!
 //! This library is not meant to be used in a high performance scenario. It uses synchronous IO.
-//! 
+//!
 //! ## Documentation
-//! 
+//!
 //! Documenatation can be found at [docs.rs](https://docs.rs/nakadion)
-//! 
+//!
 //! ## License
-//! 
-//! Nakadion is distributed under the terms of both the MIT license and the Apache License (Version 2.0).
-//! 
+//!
+//! Nakadion is distributed under the terms of both the MIT license
+//! and the Apache License (Version 2.0).
+//!
 //! See LICENSE-APACHE and LICENSE-MIT for details.#![recursion_limit = "1024"]
 #[macro_use]
 extern crate log;
@@ -155,14 +158,14 @@ pub struct EventTypeInfo {
 }
 
 impl EventTypeInfo {
-    /// Returns the number of partitions this `EventType` is 
+    /// Returns the number of partitions this `EventType` is
     /// published over.
     pub fn num_partitions(&self) -> usize {
         self.partitions.len()
     }
 }
 
-/// A stream can provide multiple `EventTypes` where each of them can have 
+/// A stream can provide multiple `EventTypes` where each of them can have
 /// its own partitioning setup.
 #[derive(Debug, Deserialize)]
 pub struct StreamInfo {
@@ -174,7 +177,11 @@ impl StreamInfo {
     /// Returns the number of partitions of the `EventType`
     /// that has the most partitions.
     pub fn max_partitions(&self) -> usize {
-        self.event_types.iter().map(|et| et.num_partitions()).max().unwrap_or(0)
+        self.event_types
+            .iter()
+            .map(|et| et.num_partitions())
+            .max()
+            .unwrap_or(0)
     }
 }
 
@@ -254,26 +261,29 @@ impl<C: NakadiConnector> NakadiClient<C> {
                                      handler: H,
                                      settings: WorkerSettings)
                                      -> Result<(Self, JoinHandle<()>), String> {
-        let (worker, handle) = NakadiWorker::new(connector.clone(), handler, subscription_id, settings)?;
+        let (worker, handle) =
+            NakadiWorker::new(connector.clone(), handler, subscription_id, settings)?;
         Ok((NakadiClient {
-            worker: worker,
-            connector: connector,
-        },
-         handle))
+                worker: worker,
+                connector: connector,
+            },
+            handle))
     }
 
-    /// Configure the client from environment variables. 
+    /// Configure the client from environment variables.
     ///
     /// The `SubscriptionId` is provided manually. Useful if you want to consume
     /// multiple subscriptions but want to use the same settings everywhere else.
-    pub fn from_env_with_subscription<H: Handler + 'static, T: ProvidesToken>(
-        subscription_id: SubscriptionId,
-        handler: H,
-        token_provider: T) -> Result<(NakadiClient<HyperClientConnector>, JoinHandle<()>), String> {
+    pub fn from_env_with_subscription<H: Handler + 'static, T: ProvidesToken>
+        (subscription_id: SubscriptionId,
+         handler: H,
+         token_provider: T)
+         -> Result<(NakadiClient<HyperClientConnector>, JoinHandle<()>), String> {
         let connector = HyperClientConnector::from_env(Box::new(token_provider))?;
         let connector = Arc::new(connector);
         let worker_settings = WorkerSettings::from_env()?;
-        let (worker, handle) = NakadiWorker::new(connector.clone(), handler, subscription_id, worker_settings)?;
+        let (worker, handle) =
+            NakadiWorker::new(connector.clone(), handler, subscription_id, worker_settings)?;
         let client: NakadiClient<HyperClientConnector> = NakadiClient {
             worker: worker,
             connector: connector,
@@ -282,16 +292,26 @@ impl<C: NakadiConnector> NakadiClient<C> {
 
     }
 
-    /// Configure the client solely from environment variables. 
+    /// Configure the client solely from environment variables.
     /// Including the `SubscriptionId`.
-    pub fn from_env<H: Handler + 'static, T: ProvidesToken>(
-        handler: H,
-        token_provider: T) -> Result<(NakadiClient<HyperClientConnector>, JoinHandle<()>), String> {
+    pub fn from_env<H: Handler + 'static, T: ProvidesToken>
+        (handler: H,
+         token_provider: T)
+         -> Result<(NakadiClient<HyperClientConnector>, JoinHandle<()>), String> {
         let subscription_id: SubscriptionId = match env::var("NAKADION_SUBSCRIPTION_ID") {
-            Ok(env_val) => SubscriptionId(env_val.parse().map_err(|err| format!("Could not parse 'NAKADION_SUBSCRIPTION_ID': {}", err))?),
-            Err(err) => return Err(format!("Could not get env var 'NAKADION_SUBSCRIPTION_ID': {}", err))
+            Ok(env_val) => SubscriptionId(
+                env_val
+                    .parse()
+                    .map_err(|err| format!(
+                        "Could not parse 'NAKADION_SUBSCRIPTION_ID': {}",
+                        err))?),
+            Err(err) => {
+                return Err(format!("Could not get env var 'NAKADION_SUBSCRIPTION_ID': {}", err))
+            }
         };
-        NakadiClient::<HyperClientConnector>::from_env_with_subscription(subscription_id, handler, token_provider)
+        NakadiClient::<HyperClientConnector>::from_env_with_subscription(subscription_id,
+                                                                         handler,
+                                                                         token_provider)
     }
 
     /// Get access to the underlying `NakadiConnector`.
